auth-connector/requirements.txt
auth-connector>=1.0.0
requests>=2.28.0
PyJWT>=2.4.0
flask>=2.0.0